+++ spring-social/src/main/java/com/bodhpsychometric/dto/MqtScoreResponse.java
@@ -1,8 +1,8 @@
 package com.bodhpsychometric.dto;
 
-/** One MQT scoring entry as the API returns it — name included for display. */
+/** One MQT scoring entry as the API returns it — name included for display. Decimal since V25. */
 public record MqtScoreResponse(
         Long measuredQualityTypeId,
         String measuredQualityTypeName,
-        int score) {
+        double score) {
 }
